requirements.txt
fastmcp>=0.1.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.2.0
langchain-anthropic>=0.3.0
python-dotenv>=1.0.0
